Add type casting to improve memory performance

+++ src/CurrentGMT.java
@@ -1,12 +1,12 @@
 public class CurrentGMT {
     public static void main(String[] args) {
         long currentTime = System.currentTimeMillis();
-        long totalSeconds = currentTime / 1000;
-        long currentSecond = totalSeconds % 60;
-        long totalMinutes = totalSeconds / 60;
-        long currentMinute = totalMinutes % 60;
-        long totalHours = totalMinutes / 60;
-        long currentHour = totalHours % 24;
+        int totalSeconds = (int)(currentTime / 1000);
+        byte currentSecond = (byte)(totalSeconds % 60);
+        int totalMinutes = totalSeconds / 60;
+        byte currentMinute = (byte)(totalMinutes % 60);
+        int totalHours = totalMinutes / 60;
+        byte currentHour = (byte)(totalHours % 24);
         System.out.println("The current time is: " + currentHour + ":" +
                 currentMinute + ":" + currentSecond + " GMT.");
     }
